requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-caching==2.1.0
boto3==1.34.34
botocore==1.34.34
python-jose==3.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
yfinance==0.2.55
feedparser==6.0.10
xmltodict==0.13.0
praw==7.7.1
textblob==0.17.1
pandas==2.1.3
numpy==1.26.2
jwt
chromadb
sentence-transformers